moving entity to the right package
